move logo to mode 0
This looks just the same (for now) but uses about 60% of the screen
memory.  This is good, because we'll soon be quadrupling the memory usage.

Specifically, we use a 61-character narrow screen mode, and only store
12 lines instead of 16 (plus a single blank line) for a total of 5856
bytes per frame instead of 10240.  Of course, we only actually *care*
about 1/8 of those bytes.